joint_trajectory_controller: resolve implicit cast errors

+++ joint_trajectory_controller/src/trajectory.cpp
@@ -174,8 +174,8 @@ bool Trajectory::sample(
 
         interpolate_between_points(t0, point, t1, next_point, sample_time, output_state);
       }
-      start_segment_itr = begin() + i;
-      end_segment_itr = begin() + (i + 1);
+      start_segment_itr = begin() + static_cast<long>(i);
+      end_segment_itr = begin() + static_cast<long>((i + 1));
       last_sample_idx_ = i;
       return true;
     }
